Changed record's side length type from int to float, started to work on the sequential file

+++ src/main/java/org/example/comparators/RecordSideAComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class RecordSideAComparator implements Comparator<Record> {
     @Override
     public int compare(Record a,Record b){
-        return Integer.compare(a.getA(),b.getA());
+        return Float.compare(a.getA(),b.getA());
     }
 }
